+++ modules/School_Setup/Schools.php
@@ -1,6 +1,7 @@
 <?php
 
 require_once 'ProgramFunctions/Fields.fnc.php';
+require_once 'ProgramFunctions/StudentsUsersInfo.fnc.php';
 
 DrawHeader( ProgramTitle() );
 
@@ -54,16 +55,39 @@ if ( $_REQUEST['modfunc']=='update')
 
 					foreach ( (array) $_REQUEST['values'] as $column => $value)
 					{
-						if (1)//!empty($value) || $value=='0')
+						if ( ! is_array( $value ) )
 						{
 							//FJ check numeric fields
 							if ( $fields_RET[str_replace('CUSTOM_','',$column)][1]['TYPE'] == 'numeric' && $value!='' && !is_numeric($value))
 							{
-								$error[] = _('Please enter valid Numeric data.');
+								$error[] = _( 'Please enter valid Numeric data.' );
 								continue;
 							}
 
-							$sql .= $column."='".$value."',";
+							$sql .= $column . "='" . $value . "',";
+							$go = true;
+						}
+						else
+						{
+							// Select multiple from options.
+							// FJ fix bug none selected not saved.
+							$sql_multiple_input = '';
+
+							foreach ( (array) $value as $val )
+							{
+								if ( $val )
+								{
+									$sql_multiple_input .= $val . '||';
+								}
+							}
+
+							if ( $sql_multiple_input )
+							{
+								$sql_multiple_input = "||" . $sql_multiple_input;
+							}
+
+							$sql .= $column . "='" . $sql_multiple_input . "',";
+
 							$go = true;
 						}
 					}
@@ -296,25 +320,30 @@ if ( ! $_REQUEST['modfunc'] )
 		) . '</td></tr>';
 	}
 
-	//FJ add School Fields
-	$fields_RET = DBGet(DBQuery("SELECT ID,TITLE,TYPE,DEFAULT_SELECTION,REQUIRED FROM SCHOOL_FIELDS ORDER BY SORT_ORDER,TITLE"));
-	$fields_RET = ParseMLArray($fields_RET,'TITLE');
+	// FJ add School Fields.
+	$fields_RET = DBGet( DBQuery( "SELECT ID,TITLE,TYPE,SELECT_OPTIONS,DEFAULT_SELECTION,REQUIRED
+		FROM SCHOOL_FIELDS
+		ORDER BY SORT_ORDER,TITLE" ) );
+
+	$fields_RET = ParseMLArray( $fields_RET, 'TITLE' );
 
 	if ( count( $fields_RET ) )
+	{
 		echo '<tr><td colspan="3"><hr /></td></tr>';
+	}
 
 	foreach ( (array) $fields_RET as $field )
 	{
 		$value_custom = '';
 
-		if ( $_REQUEST['new_school'] != 'true' )
+		if ( $_REQUEST['new_school'] !== 'true' )
 		{
 			$value_custom = DBGet( DBQuery( "SELECT CUSTOM_" . $field['ID'] . "
 				FROM SCHOOLS
 				WHERE ID='" . UserSchool() . "'
 				AND SYEAR='" . UserSyear() . "'" ) );
 
-			$value_custom = $value_custom[1]['CUSTOM_' . $field['ID']];
+			$value_custom = $value_custom[1][ 'CUSTOM_' . $field['ID'] ];
 
 			$div = true;
 		}
@@ -338,7 +367,7 @@ if ( ! $_REQUEST['modfunc'] )
 					$value_custom,
 					'values[CUSTOM_' . $field['ID'] . ']',
 					$title_custom,
-					( $field['REQUIRED'] ? ' required' : '' ),
+					'maxlength=255' . ( $field['REQUIRED'] ? ' required' : '' ),
 					$div
 				);
 
@@ -372,6 +401,139 @@ if ( ! $_REQUEST['modfunc'] )
 					$value_custom,
 					'values[CUSTOM_' . $field['ID'] . ']',
 					$title_custom,
+					'maxlength=5000' . ( $field['REQUIRED'] ? ' required' : '' ),
+					$div
+				);
+
+				break;
+
+			// Add School Field types.
+			case 'radio':
+				echo CheckboxInput(
+					$value_custom,
+					'values[CUSTOM_' . $field['ID'] . ']',
+					$title_custom,
+					'',
+					false,
+					'Yes',
+					'No',
+					$div,
+					( $field['REQUIRED'] ? ' required' : '' )
+				);
+
+				break;
+
+			case 'multiple':
+				// Global.
+				$value[ 'CUSTOM_' . $field['ID'] ] = $value_custom;
+
+				echo _makeMultipleInput( 'CUSTOM_' . $field['ID'], $title_custom, 'values' );
+
+				break;
+
+			case 'select':
+			case 'autos':
+			case 'edits':
+			case 'codeds':
+			case 'exports':
+				$options = $select_options = array();
+
+				$col_name = 'CUSTOM_' . $field['ID'];
+
+				if ( $field['SELECT_OPTIONS'] )
+				{
+					$options = explode(
+						"\r",
+						str_replace( array( "\r\n", "\n" ), "\r", $field['SELECT_OPTIONS'] )
+					);
+				}
+
+				foreach ( (array) $options as $option )
+				{
+					$value = $option;
+
+					// Exports specificities.
+					if ( $field['TYPE'] === 'exports' )
+					{
+						$option = explode( '|', $option );
+
+						$option = $value = $option[0];
+					}
+					// Codeds specificities.
+					elseif ( $field['TYPE'] === 'codeds' )
+					{
+						list( $value, $option ) = explode( '|', $option );
+					}
+
+					if ( $value !== ''
+						&& $option !== '' )
+					{
+						$select_options[ $value ] = $option;
+					}
+				}
+
+				// Get autos / edits pull-down edited options.
+				if ( $field['TYPE'] === 'autos'
+					|| $field['TYPE'] === 'edits' )
+				{
+					if ( $value_custom === '---'
+						|| count( $select_options ) <= 1 )
+					{
+						// FJ new option.
+						echo TextInput(
+							$value_custom === '---' ?
+								array( '---', '<span style="color:red">-' . _( 'Edit' ) . '-</span>' ) :
+								$value_custom,
+							'values[CUSTOM_' . $field['ID'] . ']',
+							$title_custom,
+							( $field['REQUIRED'] === 'Y' ? 'required' : '' ),
+							$div
+						);
+
+						break;
+					}
+
+					$sql_options = "SELECT DISTINCT s." . $col_name . ",upper(s." . $col_name . ") AS SORT_KEY
+						FROM SCHOOLS s
+						WHERE (s.SYEAR='" . UserSyear() . "' OR s.SYEAR='" . ( UserSyear() - 1 ) . "')
+						AND s." . $col_name . " IS NOT NULL
+						AND s." . $col_name . " != ''
+						ORDER BY SORT_KEY";
+
+					$options_RET = DBGet( DBQuery( $sql_options ) );
+
+					// Add the 'new' option, is also the separator.
+					$select_options['---'] = '-' . _( 'Edit' ) . '-';
+
+					foreach ( (array) $options_RET as $option )
+					{
+						$option_value = $option[ $col_name ];
+
+						if ( ! isset( $select_options[ $option_value ] ) )
+						{
+							$select_options[ $option_value ] = '<span style="color:blue">' .
+								$option_value . '</span>';
+						}
+					}
+
+					// Make sure the current value is in the list.
+					if ( $value_custom != ''
+						&& ! isset( $select_options[ $value_custom ] ) )
+					{
+						$select_options[ $value_custom ] = array(
+							$value_custom,
+							'<span style="color:' . ( $field['TYPE'] === 'autos' ? 'blue' : 'green' ) . '">' .
+								$value_custom . '</span>'
+						);
+					}
+				}
+
+				echo SelectInput(
+					$value_custom,
+					'values[CUSTOM_' . $field['ID'] . ']',
+					$title_custom,
+					$select_options,
+					'N/A',
 					( $field['REQUIRED'] ? ' required' : '' ),
 					$div
 				);